move IrcServerListing to a separate file
