+++ src/service/irc/IrcService.cpp
@@ -7,6 +7,7 @@
 #include "event/EventQuit.hpp"
 #include "event/EventQuery.hpp"
 #include "event/irc/IrcChannelListing.hpp"
+#include "event/irc/IrcServerListing.hpp"
 #include "event/irc/EventIrcActivateService.hpp"
 #include "event/irc/EventIrcChatListing.hpp"
 #include "event/irc/EventIrcSettingsListing.hpp"
